Split vagrant directory into ec2 and libvirt directories
Vagrant will always read the the entire Vagrantfile - even those parts which are
not related to the provider being used. This means that any undefined variable
mentioned in the Vagrantfile will cause Vagrant to error out. Having to define
dummy values for, say, libvirt when one is only using ec2 is silly, so splitting
the configuration into two separate Vagrantfiles was the only reasonable option.
As the Vagrantfiles share very little content, the amount of redundancy is
still manageable.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>